@@ -1,10 +1,9 @@
 """Tibber home"""
-from __future__ import annotations
-
 import asyncio
 import datetime as dt
 import logging
-from typing import TYPE_CHECKING, Callable
+from typing import TYPE_CHECKING
+from collections.abc import Callable
 
 from dateutil.parser import parse
 from gql import gql
